# openapi/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/app-management": {
      "post": {
        "operationId": "AppManagementController_createApp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "App created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new app",
        "tags": [
          "app-management"
        ]
      },
      "get": {
        "operationId": "AppManagementController_getApps",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Apps retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all apps",
        "tags": [
          "app-management"
        ]
      }
    },
    "/api/app-management/{id}": {
      "get": {
        "operationId": "AppManagementController_getApp",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "App ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "App retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "App not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get an app by ID",
        "tags": [
          "app-management"
        ]
      },
      "put": {
        "operationId": "AppManagementController_updateApp",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "App ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "App updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "App not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an app",
        "tags": [
          "app-management"
        ]
      },
      "delete": {
        "operationId": "AppManagementController_deleteApp",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "App ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "App deleted successfully"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "App not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an app",
        "tags": [
          "app-management"
        ]
      }
    },
    "/api/app-management/{id}/statistics": {
      "get": {
        "operationId": "AppManagementController_getAppStatistics",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "App ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "App statistics retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get app statistics",
        "tags": [
          "app-management"
        ]
      }
    },
...
    "/api/privacy-policies/app/{appId}": {
      "get": {
        "operationId": "PrivacyPoliciesController_getPolicyByAppId",
        "parameters": [
          {
            "name": "appId",
            "required": true,
            "in": "path",
            "description": "Application ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Privacy policy found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyPolicyDto"
                }
              }
            }
          },
          "404": {
            "description": "Privacy policy not found"
          }
        },
        "summary": "Get privacy policy by app ID",
        "tags": [
          "privacy-policies"
        ]
      }
    },
...
    "/api/privacy-policies/web/app/{appId}": {
      "get": {
        "operationId": "PrivacyPoliciesController_renderPolicyPageByAppId",
        "parameters": [
          {
            "name": "appId",
            "required": true,
            "in": "path",
            "description": "Application ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Render privacy policy page by app ID",
        "tags": [
          "privacy-policies"
        ]
      }
    },
...
    "/api/privacy-policy/app/{appId}": {
      "get": {
        "operationId": "PublicPrivacyPoliciesController_renderPolicyPageByAppId",
        "parameters": [
          {
            "name": "appId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublicPrivacyPolicies"
        ]
      }
    },
...
  },
  "info": {
    "title": "Doctopus API",
    "description": "Modern NestJS API with comprehensive features",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    },
    {
      "url": "https://tree2u.kobotogether.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateAppDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Unique app name (used as identifier)",
            "example": "my-new-app",
            "minLength": 3,
            "maxLength": 50
          },
          "displayName": {
            "type": "string",
            "description": "Human-readable display name",
            "example": "My New App",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "App description",
            "example": "A comprehensive app for managing containers",
            "maxLength": 500
          },
          "domain": {
            "type": "string",
            "description": "App domain (optional)",
            "example": "myapp.example.com"
          },
          "settings": {
            "type": "object",
            "description": "App-specific settings (JSON object)",
            "example": {
              "theme": "dark",
              "language": "en"
            }
          }
        },
        "required": [
          "name",
          "displayName"
        ]
      },
      "AppResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "App ID",
            "example": "uuid-string"
          },
          "name": {
            "type": "string",
            "description": "App name",
            "example": "my-app"
          },
          "displayName": {
            "type": "string",
            "description": "Display name",
            "example": "My App"
          },
          "description": {
            "type": "string",
            "description": "App description",
            "example": "A comprehensive app for managing containers"
          },
          "domain": {
            "type": "string",
            "description": "App domain",
            "example": "myapp.example.com"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the app is active",
            "example": true
          },
          "settings": {
            "type": "object",
            "description": "App-specific settings",
            "example": {
              "theme": "dark",
              "language": "en"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp",
            "example": "2023-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "displayName",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateAppDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "description": "Human-readable display name",
            "example": "Updated App Name",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "App description",
            "example": "Updated app description",
            "maxLength": 500
          },
          "domain": {
            "type": "string",
            "description": "App domain (optional)",
            "example": "updated.example.com"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the app is active",
            "example": true
          },
          "settings": {
            "type": "object",
            "description": "App-specific settings (JSON object)",
            "example": {
              "theme": "light",
              "language": "es"
            }
          }
        }
      },
      "PrivacyPolicyDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}